caso_practico_2 funciona todo, solo falta hacer la tabla que pide el ejercicio

+++ caso_practico_2/templates/fb_db_vm.init.sh
@@ -1,20 +1,60 @@
 #!/bin/bash
 
+LOG_FILE="/home/ubuntu/vm_db.log"
+
+log() {
+    echo "$(date +'%Y-%m-%d %H:%M:%S') - $1" | tee -a $LOG_FILE
+}
+
+log "Starting script"
+
+log "Updating apt repository"
+sudo apt-get update
+
+log 'Changing binding address'
 sudo sed -i '/bind-address/s/127.0.0.1/0.0.0.0/' /etc/mysql/mysql.conf.d/mysqld.cnf
 
-sudo mysql << EOF
-	CREATE USER '${db_user}'@'%' IDENTIFIED BY '${db_pass}';
-	CREATE DATABASE IF NOT EXISTS ${db_name};
-	GRANT ALL PRIVILEGES ON metabasedb.* TO '${db_user}'@'%';
-	FLUSH PRIVILEGES;
+log 'Creating User'
+if sudo mysql << EOF
+    CREATE USER '${db_user}'@'%' IDENTIFIED BY '${db_pass}';
+    CREATE DATABASE IF NOT EXISTS ${db_name};
+    GRANT ALL PRIVILEGES ON ${db_name}.* TO '${db_user}'@'%';
+    FLUSH PRIVILEGES;
 EOF
+then
+    log "User and database created successfully"
+else
+    log "Failed to create user and database"
+fi
 
-cd tmp/
+cd /tmp
 
-curl -LJO https://github.com/FrancoBertoldiMariglio/Teleinformatica/raw/main/caso_practico_2/google-mobility.sql.gz -o /tmp/google-mobility.sql.gz
+log 'Downloading SQL file'
+if curl -LJO https://github.com/FrancoBertoldiMariglio/Teleinformatica/raw/main/caso_practico_2/google-mobility.sql.gz -o google-mobility.sql.gz; then
+    log "SQL file downloaded successfully"
+else
+    log "Failed to download SQL file"
+fi
 
-gzip -d /tmp/google-mobility.sql.gz
+log 'Decompressing SQL file'
+sudo gzip -d google-mobility.sql.gz
 
-sudo mysql ${db_name} < /tmp/google-mobility.sql
+if [[ ! -f google-mobility.sql ]]; then
+	log "Failed to decompress SQL file"
+	exit 1
+fi
 
-rm /tmp/google-mobility.sql 
+log 'Importing SQL file'
+sudo mysql ${db_name} < google-mobility.sql
+
+row_count=$(sudo mysql ${db_name} -e "SELECT COUNT(*) FROM mobility;" | tail -n 1)
+if [[ $row_count -gt 0 ]]; then
+    log "Data correctly inserted into the database"
+else
+    log "Error inserting data into the database"
+fi
+
+log 'Removing SQL file'
+rm /tmp/google-mobility.sql
+
+log "Script finished succesfully"
